Update device.py
Missing fusegroup added
Change power_get(): Finally this calculates the power which goes in  or out of the device. You can use outputHomePower and gridInputPower, or add all incoming or outgoing power like solarInputPower, packInputPower, outputPackPower, gridOffPower and so on.
This solution doesn't care on single values, just get the sum of in- or outgoing power. 
This need to remove the additional subtract of gridOffPower for SF 800 pro and SF 2400AC